Fixed another typo

+++ sources/hopper/grid.h
@@ -5,7 +5,7 @@ struct linkedlist {
 	particle_t *data;
 } linkedlist;
 
-void grid_init(int size);
+void grid_init(double size);
 void grid_add(particle_t *particle);
 void grid_remove(particle_t *particle);
 linkedlist_t* grid_get_collisions(particle_t *particle);
